+++ devices/MIMX8MM5/drivers/fsl_iomuxc.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2021 NXP
+ * Copyright 2018-2020 NXP
  * All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
@@ -606,11 +606,11 @@ extern "C" {
  * @param configRegister The config register_
  * @param inputOnfield   The pad->module input inversion_
  */
-static inline void IOMUXC_SetPinMux(uintptr_t muxRegister,
+static inline void IOMUXC_SetPinMux(uint32_t muxRegister,
                                     uint32_t muxMode,
-                                    uintptr_t inputRegister,
+                                    uint32_t inputRegister,
                                     uint32_t inputDaisy,
-                                    uintptr_t configRegister,
+                                    uint32_t configRegister,
                                     uint32_t inputOnfield)
 {
     *((volatile uint32_t *)muxRegister) =
@@ -637,11 +637,11 @@ static inline void IOMUXC_SetPinMux(uintptr_t muxRegister,
  * @param configRegister The config register_
  * @param configValue    The pin config value_
  */
-static inline void IOMUXC_SetPinConfig(uintptr_t muxRegister,
+static inline void IOMUXC_SetPinConfig(uint32_t muxRegister,
                                        uint32_t muxMode,
-                                       uintptr_t inputRegister,
+                                       uint32_t inputRegister,
                                        uint32_t inputDaisy,
-                                       uintptr_t configRegister,
+                                       uint32_t configRegister,
                                        uint32_t configValue)
 {
     if (configRegister)
